Refactoringand minor fixes
